tui: tidy dangling interfaces
* Post-refactor clean up of dangling code.
